backend/requirements.txt
fastapi
uvicorn
redis
python-dotenv
langchain
langchain-google-genai
pydantic
sqlalchemy
psycopg2-binary
pandas
xgboost
scikit-learn
faiss-cpu
numpy
python-multipart
jose[cryptography]
passlib[bcrypt]
websockets
